chore: change unnecessary int64 to int32 (#9108)
* chore: change unnecessary int64 to int32

* update

Former-commit-id: fb59c0f38a6886f8033ddfc0eb74395756db63ce

+++ backend/api/v1/database_service.go
@@ -1600,7 +1600,7 @@ func (s *DatabaseService) checkDatabasePermission(ctx context.Context, projectID
 
 type totalValue struct {
 	totalQueryTime time.Duration
-	totalCount     int64
+	totalCount     int32
 }
 
 // ListSlowQueries lists the slow queries.
